# Create backend-agnostic Backup class