Fixes #3201

The new file `backup_general.py` hosts a backend-agnostic Backup class
that will query through `aiida.orm.QueryBuilder`, instead of using the
backend specific query language.
This makes the backup-module completely backend-agnostic, drastically
simplifying it, however, it may also become slightly slower.
A mid-way solution should be considered, where perhaps one may employ
backend-specific Backup classes, performing backend-specific queries,
increasing speed - but only if the backup is otherwise exceedingly slow.
It is recommended to be as agnostic as possible.

It is recommended to do benchmark tests of the speed of this change before merging.

This PR is derived from PR #3205