@CasperWA 2019-08-29
It should be noted that anyone interested is welcome to perform benchmark tests of this :)