@CasperWA 2019-10-03
I have done some testing.

Creating two new databases using the different backends, they have been filled with the same Nodes.
Each Node is an `Int(1)` and has a 1 kB `extra` binary string.
I have filled up the databases with 100 MB worth of 1 kB Nodes, amounting to 102400 Nodes (When 1 MB = 1024 kB).

For each backup run, the `backup_info.json` has its key `oldest_object_backedup` reset to the value `'2019-01-01 00:00:00+00:00'`, so that a backup run can be run in a for loop for N times, where N can be varied, and all Nodes are backed up for each run.

The following table shows the number of backup runs performed, as well as the average time for the runs. Furthermore, 'before' corresponds to the develop branch as is (with the split Backup classes for each backend) and 'after' then corresponds to the changes introduced in this PR.

|   | Django | SQLAlchemy |
| --- | --- | --- |
| before | 100 runs, **0.66s** | 100 runs, **14s** |
| after | 110 runs, **1.6s** | 105 runs, **18s** |

The reason for the different number of runs is due to initial testing of the scripts used.

Below is a graph show all timings for the runs:
![backup_times](https://user-images.githubusercontent.com/43357585/66149580-ad6ed600-e613-11e9-97b9-6e84bb782f1f.png)