@CasperWA 2019-10-03
What is missing is the timings for different sized databases, however, I am not sure it is needed?

In the end, what it shows is a decrease in efficiency of **142%** for the Django backend and of **29%** for the SQLAlchemy backend.

This sounds bad, but in absolute time the increase for a database of 100k+ Nodes is not severe, I think.
The gain is a more neat Backup class hierarchy and one less place in the code, where the use of direct backend querying is avoided.