@giovannipizzi 2019-10-03
Thanks for the accurate testing! Do you know why SQLA is so much slower?