@sphuber 2019-10-04
I looked through the code and for me this is good to merge. I'm ok with the decrease in performance for reduced code complexity/fragility. If performance becomes a problem, we can address it in the future