@CasperWA 2019-10-04
> Thanks for the accurate testing! Do you know why SQLA is so much slower?

No. Actually, it quite surprised me, since I believe we have a bit more steps to go through when querying the Django backend, versus the SQLAlchemy backend through the QueryBuilder?

For completion, (and this has nothing to do with the question above) it is worth mentioning that this testing is done only for the difference introduced in this PR to how the database is queried. Therefore, while the Backup classes indeed also copies the repository, it does not the target of the current investigation. Therefore, all Nodes have empty repository folders and the "meat" of the Nodes have been placed in such a way as to fill the database, as well as create a multitude of Nodes - making querying more taxing.